blackjack
Blackjack is a simple, popular card game that is played in many casinos.
Cards in Blackjack have the following values: an ace may be valued as
either 1 or 11 (player's choice), face cards (kings, queens and jacks)
are valued at 10 and the value of the remaining cards corresponds to
their number. During a round of Blackjack, the players plays against a
dealer with the goal of building a hand (a collection of cards) whose
cards have a total value that is higher than the value of the dealer's
hand, but not over 21.  (A round of Blackjack is also sometimes referred
to as a hand.)