RiceRocks (Asteroids)
For our last mini-project, we will complete the implementation of
RiceRocks, an updated version of Asteroids,  that we began last week.
You may start with either your code or the program template which
includes a full implementation of Spaceship and will be released
immediately after the deadline for the Spaceship mini-project (by making
the preceding link live).  If you start with your own code, you should
add the splash screen image that you dismiss with a mouse click before
starting this mini-project.  We strongly recommend using Chrome for this
mini-project since Chrome's superior performance will become apparent
when your program attempts to draw dozens of sprites.